Re-organize validator client CLI
