+++ packages/lodestar-cli/src/cmds/account/paths.ts
@@ -7,11 +7,6 @@ export interface IAccountPaths {
   walletsDir: string;
 }
 
-export interface IValidatorPaths {
-  validatorsDbDir: string;
-  validatorDbDir: (pubkey: string) => string;
-}
-
 /**
  * Defines the path structure of the account files
  * 
@@ -49,28 +44,6 @@ export function getAccountPaths(options: Partial<IAccountPaths> & Pick<IGlobalAr
 }
 
 /**
- * Defines the path structure of the validator files
- * 
- * ```bash
- * $validatorRootDir
- * └── validator-db
- *     └── 0x8e41b969493454318c27ec6fac90645769331c07ebc8db5037...
- *         └── (db files)
- * ```
- */
-export function getValidatorPaths(options: Partial<IValidatorPaths> & Pick<IGlobalArgs, "rootDir">): IValidatorPaths {
-  const rootDir = options.rootDir || ".lodestart";
-  const validatorsDbDir = path.join(rootDir, options.validatorsDbDir || "validator-db");
-  return {
-    validatorsDbDir,
-    validatorDbDir: (pubkey: string) => path.join(validatorsDbDir, pubkey)
-  };
-}
-
-/**
  * Constructs representations of the path structure to show in command's description
  */
-export const defaultPaths = {
-  ...getAccountPaths({rootDir: "$rootDir"}),
-  ...getValidatorPaths({rootDir: "$rootDir"})
-};
+export const defaultAccountPaths = getAccountPaths({rootDir: "$rootDir"});
